requirements.txt
opencv-python
mediapipe
tensorflow-io==0.23.1
numpy
pandas
matplotlib